+++ src/renderer/components/observationsHelpers.tsx
